+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - Change Highlight Symbol to actually be drawn (fixes #1)
   - The Highlight Symbol draw width is configurable
   - The Highlight Symbol draw behavior is configurable between "Indent only selected"(`CombineIndent`) and "indent all"(`Static`)
+- Add ability to set style for `widget::Indicator` and `widget::OrIndicators` (fixes #2)
+- Apply common style for indent area
 
 ## 0.1.0